Allow optional parameters for submit
`qat.qlmaas.qpus.QLMaaSQPU.submit()` admits several parameters to
override the resources required to run the job. This may be necessary
for QPUs such as AnalogQPU which may not estimate correctly the required
memory. Since pulser_myqlm could submit to different types of QPUs,
these parameters are not passed directly as arguments of `submit()` but
as a dict.